requirements.txt

google-generativeai
streamlit
dotenv
streamlit_mic_recorder
pytube
streamlit_player
pillow